backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
httpx
playwright
weasyprint
openai
python-multipart
celery
redis
jinja2
numpy
statistics
passlib[bcrypt]
python-jose[cryptography]
beautifulsoup4
user-agents
